pyproject.toml
[project]
name = "apartment-finder"
version = "0.1.0"
description = "Automated rental listing aggregator with AI-powered extraction"
requires-python = ">=3.11"
dependencies = [
    "crawlee[playwright]>=0.6.0",
    "camoufox>=0.4.0",
    "zendriver>=0.3.0",
    "anthropic>=0.42.0",
    "geopy>=2.4.0",
    "pyyaml>=6.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
